requirements.txt
streamlit
torch
torchvision
numpy
Pillow
scikit-image
tqdm
timm